feat(refactor): 전체 피드백 내용 적용
**
강사님의 피드백을 반영하여 JCF(Java Collections Framework)라는
   구체적인 구현 기술이 명시되었지만 그 이상의 기능을 가져 표준 구현체 클래스 이름인 Impl로 변경합니다

멘토님의 피드백을 반영하여 기존 정적 팩토리 메서드에서 protected -> private 로 접근 제한을 수정합니다

NPE예방을 위해 ChannelMessageRepositoryImpl의 findAllById 메서드의 equals비교 순서를 역순으로 변경합니다.

nullcheck없는 부분에 nullcheck를 추가합니다

Optional의 isPresent 메서드를 성능 문제로 orElseThrow || orElse || ifPresentOrElse 메서드로 변경하였습니다.

기존 AppConfig를 AppConfigRegacy로 변경하고 더 이상 사용 되지 않습니다.
**
