+++ src/test/java/com/sprint/mission/discodeit/JCFUserServiceTest.java
@@ -5,7 +5,7 @@ import com.sprint.mission.discodeit.repository.impl.ChannelMessageRepositoryImpl
 import com.sprint.mission.discodeit.repository.impl.DirectMessageRepositoryImpl;
 import com.sprint.mission.discodeit.repository.impl.ParticipationRepositoryImpl;
 import com.sprint.mission.discodeit.repository.impl.UserRepositoryImpl;
-import com.sprint.mission.discodeit.service.jcf.UserServiceImpl;
+import com.sprint.mission.discodeit.service.impl.UserServiceImpl;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
